Refactured net generation dialog into multiple tabs for better data visualization and future flexibility to integrate multiple heat generators
