+++ src/districtheatingsim/gui/CalculationTab/timeseries_dialog.py
@@ -1,0 +1,134 @@
+"""
+Filename: calculation_dialogs.py
+Author: Dipl.-Ing. (FH) Jonas Pfeiffer
+Date: 2025-02-11
+Description: Contains the Dialogs for the CalculationTab.
+"""
+
+import os
+import sys
+
+from PyQt5.QtWidgets import QVBoxLayout, QLineEdit, QLabel, QDialog,  \
+    QPushButton, QHBoxLayout, QFileDialog, QMessageBox
+
+def get_resource_path(relative_path):
+    """ Get the absolute path to the resource, works for dev and for PyInstaller """
+    if getattr(sys, 'frozen', False):
+        # Wenn die Anwendung eingefroren ist, ist der Basispfad der Temp-Ordner, wo PyInstaller alles extrahiert
+        base_path = sys._MEIPASS
+    else:
+        # Wenn die Anwendung nicht eingefroren ist, ist der Basispfad der Ordner, in dem die Hauptdatei liegt
+        base_path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
+    return os.path.join(base_path, relative_path)
+
+class TimeSeriesCalculationDialog(QDialog):
+    """
+    Dialog for time series calculation based on user inputs.
+    
+    Attributes:
+        base_path (str): Base path for file dialogs.
+        parent (QWidget): Parent widget.
+    """
+
+    def __init__(self, base_path, parent=None):
+        super().__init__(parent)
+        self.base_path = base_path
+        self.parent = parent
+        self.initUI()
+
+    def initUI(self):
+        """
+        Initializes the user interface for the dialog.
+        """
+        self.setWindowTitle("Zeitreihenrechnung")
+        self.resize(400, 200)
+
+        self.layout = QVBoxLayout(self)
+
+        # Zeitschritte
+        self.StartTimeStepLabel = QLabel("Zeitschritt Simulationsstart (min 0):", self)
+        self.StartTimeStepInput = QLineEdit("0", self)
+        self.EndTimeStepLabel = QLabel("Zeitschritt Simulationsende (max 8760):", self)
+        self.EndTimeStepInput = QLineEdit("8760", self)
+
+        self.layout.addWidget(self.StartTimeStepLabel)
+        self.layout.addWidget(self.StartTimeStepInput)
+        self.layout.addWidget(self.EndTimeStepLabel)
+        self.layout.addWidget(self.EndTimeStepInput)
+
+        # Dateiauswahl
+        self.fileInputlayout = QHBoxLayout(self)
+
+        self.resultsFileLabel = QLabel("Ausgabedatei Lastgang:", self)
+        self.resultsFileInput = QLineEdit(os.path.join(self.base_path, self.parent.config_manager.get_relative_path('load_profile_path')), self)
+        self.selectresultsFileButton = QPushButton('csv-Datei auswählen')
+        self.selectresultsFileButton.clicked.connect(lambda: self.selectFilename(self.resultsFileInput))
+
+        self.fileInputlayout.addWidget(self.resultsFileLabel)
+        self.fileInputlayout.addWidget(self.resultsFileInput)
+        self.fileInputlayout.addWidget(self.selectresultsFileButton)
+
+        self.layout.addLayout(self.fileInputlayout)
+
+        # Buttons
+        buttonLayout = QHBoxLayout()
+        okButton = QPushButton("OK", self)
+        cancelButton = QPushButton("Abbrechen", self)
+        
+        okButton.clicked.connect(self.onAccept)
+        cancelButton.clicked.connect(self.reject)
+        
+        buttonLayout.addWidget(okButton)
+        buttonLayout.addWidget(cancelButton)
+
+        self.layout.addLayout(buttonLayout)
+
+    def onAccept(self):
+        """
+        Validates the inputs and accepts the dialog if valid.
+        """
+        if self.validateInputs():
+            self.accept()
+
+    def validateInputs(self):
+        """
+        Validates the start and end time steps.
+
+        Returns:
+            bool: True if inputs are valid, False otherwise.
+        """
+        start = int(self.StartTimeStepInput.text())
+        end = int(self.EndTimeStepInput.text())
+        
+        if start < 0 or start > 8760 or end < 0 or end > 8760:
+            QMessageBox.warning(self, "Ungültige Eingabe", "Start- und Endzeitschritte müssen zwischen 0 und 8760 liegen.")
+            return False
+        if start > end:
+            QMessageBox.warning(self, "Ungültige Eingabe", "Der Startschritt darf nicht größer als der Endschritt sein.")
+            return False
+        return True
+
+    def selectFilename(self, lineEdit):
+        """
+        Opens a file dialog to select a file and updates the line edit with the selected file path.
+
+        Args:
+            lineEdit (QLineEdit): The line edit to update with the selected file path.
+        """
+        filename, _ = QFileDialog.getOpenFileName(self, "Datei auswählen")
+        if filename:
+            lineEdit.setText(filename)
+
+    def getValues(self):
+        """
+        Gets the values from the dialog.
+
+        Returns:
+            dict: Dictionary containing the results filename, start time step, and end time step.
+        """
+        return {
+            'results_filename': self.resultsFileInput.text(),
+            'start': int(self.StartTimeStepInput.text()),
+            'end': int(self.EndTimeStepInput.text())
+        }
